+++ bin/utils/cli-config-utils.js
@@ -2,6 +2,7 @@ const { DEFAULT_EXT, DEFAULT_TABLE_NAME } = require('./constants');
 const { resolveClientNameWithAliases } = require('../../lib/helpers');
 const fs = require('fs');
 const path = require('path');
+const escalade = require('escalade/sync');
 const tildify = require('tildify');
 const color = require('colorette');
 const argv = require('getopts')(process.argv.slice(2));
@@ -139,6 +140,36 @@ function getStubPath(configKey, env, opts) {
   return path.join(stubDirectory, stub);
 }
 
+function findUpModulePath(cwd, packageName) {
+  const modulePackagePath = escalade(cwd, (dir, names) => {
+    if (names.includes('package.json')) {
+      return 'package.json';
+    }
+    return false;
+  });
+  try {
+    const modulePackage = require(modulePackagePath);
+    if (modulePackage.name === packageName) {
+      return path.join(
+        path.dirname(modulePackagePath),
+        modulePackage.main || 'index.js'
+      );
+    }
+  } catch (e) {}
+}
+
+function findUpConfig(cwd, name, extensions) {
+  return escalade(cwd, (dir, names) => {
+    for (const ext of extensions) {
+      const filename = `${name}.${ext}`;
+      if (names.includes(filename)) {
+        return filename;
+      }
+    }
+    return false;
+  });
+}
+
 module.exports = {
   mkConfigObj,
   resolveEnvironmentConfig,
@@ -148,4 +179,6 @@ module.exports = {
   getSeedExtension,
   getMigrationExtension,
   getStubPath,
+  findUpModulePath,
+  findUpConfig,
 };
